requirements.txt
agency-swarm==0.2.2
PyYAML==6.0.1
requests==2.31.0
beautifulsoup4==4.12.3
agency-swarm==0.2.2
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepdiff==6.7.1
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.1
fpdf==1.7.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
instructor==1.2.6
jsonref==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==1.27.0
ordered-set==4.1.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyreadline3==3.4.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.5
tenacity==8.3.0
termcolor==2.3.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.1
yarl==1.9.4
